+++ CHANGELOG.md
@@ -6,6 +6,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2026-01-30 16:36:04
+
+### Changed Files
+- `release-notes.json`
+- `src/features/dashboard/DashboardHome.tsx`
+- `src/features/policies/PolicyDashboard.tsx`
+- `src/features/policies/PolicyForm.tsx`
+- `src/features/policies/components/PolicyDialog.tsx`
+
+### Statistics
+```
+ 5 files changed, 105 insertions(+), 78 deletions(-)
+```
+
 ## [Checkpoint] - 2026-01-30 16:17:39
 
 ### Changed Files